Create Showable and remove MenuManager, menu list moved to chooser, chooser is now other class (not Menu)
