+++ Project Anolis/Assets/Scripts/Logistics/Choosers.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: ec54545f725ea5946a2d9df58014694b
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
