+++ Project Anolis/Assets/Scripts/Logistics/Choosers/Chooser.cs
@@ -1,0 +1,11 @@
+﻿using System.Collections;
+using System.Collections.Generic;
+using UnityEditor.Experimental.GraphView;
+using UnityEngine;
+
+namespace Logistics
+{
+    public abstract class Chooser : Showable
+    {
+    }
+}
