+++ Project Anolis/Assets/Scripts/Logistics/Choosers/Chooser.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 53a391f50fb7b574382f260ae5f5e9dd
+guid: ad19a4c0a559bec418eeab74c484e3f4
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
