+++ Project Anolis/Assets/Scripts/Logistics/Choosers/Implementations.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 94117ad38f8ff814fad2206f0f3e1c16
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
