+++ Project Anolis/Assets/Scripts/Logistics/MenuGroup.cs
@@ -4,16 +4,16 @@ namespace Logistics
 {
     public class MenuGroup : MonoBehaviour
     {
-        [SerializeField] private Menu _defaultMenu;
+        public Showable _defaultMenu;
 
-        public Menu CurrentMenu { get; set; }
+        public Showable CurrentMenu { get; set; }
 
         public void Awake()
         {
             ActivateMenu(_defaultMenu);
         }
 
-        public void ActivateMenu(Menu menu)
+        public void ActivateMenu(Showable menu)
         {
             CurrentMenu = menu;
             CurrentMenu.Show();
