--- Project Anolis/Assets/Scripts/Logistics/MenuManager.cs
@@ -1,16 +1,0 @@
-﻿using System.Collections.Generic;
-using UnityEngine;
-
-namespace Logistics
-{
-    public class MenuManager : MonoBehaviour
-    {
-        public Menu DefaultMenu { get; set; } = null;
-        public List<Menu> Menus { get; set; } = new List<Menu>();
-
-        public void ActivateDefault()
-        {
-            DefaultMenu.Show();
-        }
-    }
-}
