+++ Project Anolis/Assets/Scripts/Logistics/Menus/Menu.cs
@@ -1,0 +1,9 @@
+﻿using UnityEngine;
+
+namespace Logistics
+{
+    public abstract class Menu : Showable
+    {
+        public abstract bool CanHandleSelection();
+    }
+}
