+++ Project Anolis/Assets/Scripts/Logistics/Showable.cs
@@ -1,0 +1,19 @@
+﻿using UnityEngine;
+
+namespace Logistics
+{
+    public abstract class Showable : MonoBehaviour
+    {
+        [SerializeField] protected GameObject Ui;
+
+        public virtual void Show()
+        {
+            Ui.SetActive(true);
+        }
+
+        public virtual void Hide()
+        {
+            Ui.SetActive(false);
+        }
+    }
+}
