+++ Project Anolis/Assets/Scripts/Logistics/Showable.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 8a53b5fe23e67e746ba9171dcf71fc95
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
